split component definitions into multiple files
