+++ src/content_scripts/dashboard/quickCourseView/QuickCourseView.tsx
@@ -1,176 +1,12 @@
 /** @jsxImportSource preact */
+
 // @deno-types=https://raw.githubusercontent.com/preactjs/preact/10.13.2/src/index.d.ts
 import * as preact from 'preact';
 import * as hooks from 'preact/hooks';
-import { Course, RegularLectureCourse } from '../../common/storage/course.ts';
-
-type Filter = string;
-
-const semesterMap = {
-  '1/2': '前期',
-  '2/2': '後期',
-  '1/1': '通年',
-  '1/4': '第1クオーター',
-  '2/4': '第2クオーター',
-  '3/4': '第3クオーター',
-  '4/4': '第4クオーター',
-  'unfixed': '未定',
-} as const;
-const semesterOrder = {
-  '1/2': 1,
-  '2/2': 2,
-  '1/1': 3,
-  '1/4': 4,
-  '2/4': 5,
-  '3/4': 6,
-  '4/4': 7,
-  'unfixed': 8,
-} as const;
-const weekOfDayMap = {
-  'sun': '日曜',
-  'mon': '月曜',
-  'tue': '火曜',
-  'wed': '水曜',
-  'thu': '木曜',
-  'fri': '金曜',
-  'sat': '土曜',
-} as const;
-const weekOfDayOrder = {
-  'sun': 6,
-  'mon': 0,
-  'tue': 1,
-  'wed': 2,
-  'thu': 3,
-  'fri': 4,
-  'sat': 5,
-} as const;
-
-const CourseItem = (
-  props: { course: Course },
-) => {
-  const course = props.course;
-  if (course.type === 'regular-lecture') {
-    return (
-      <a
-        href={`https://cms7.ict.nitech.ac.jp/moodle40a/course/view.php?id=${course.pageId}`}
-      >
-        {weekOfDayMap[course.weekOfDay]} {course.period[0]}-{course.period[1]}限
-        {' '}
-        {course.name}
-      </a>
-    );
-  } else {
-    return (
-      <a
-        href={`https://cms7.ict.nitech.ac.jp/moodle40a/course/view.php?id=${course.pageId}`}
-      >
-        {course.name}
-      </a>
-    );
-  }
-};
-
-const ListGroup = (
-  props: { items: { key: string; element: preact.ComponentChild }[] },
-) => (
-  <ul className='list-group'>
-    {props.items.map((item) => (
-      <li
-        key={item.key}
-        className='list-group-item course-listitem border-left-0 border-right-0 border-top-0 px-2 rounded-0'
-      >
-        {item.element}
-      </li>
-    ))}
-  </ul>
-);
-
-const QuickCourseViewControl = (
-  props: {
-    filterDisplay: string;
-    filterValue: Filter;
-    setFilter: hooks.StateUpdater<Filter>;
-    filterList: { display: string; value: Filter }[];
-  },
-) => (
-  <div
-    data-region='filter'
-    className='d-flex align-items-center my-2'
-    aria-label='コースリンクコントロール'
-  >
-    <div className='dropdown mb-1 mr-auto'>
-      <button
-        type='button'
-        className='btn btn-outline-secondary dropdown-toggle'
-        data-toggle='dropdown'
-        aria-haspopup='true'
-        aria-expanded='false'
-        aria-label='グルーピングドロップダウンメニュ'
-      >
-        <span data-active-item-text>
-          {props.filterDisplay}
-        </span>
-      </button>
-      <ul
-        className='dropdown-menu'
-        role='menu'
-        data-show-active-item
-        data-skip-active-class='true'
-        data-active-item-text
-        aria-labelledby='groupingdropwodn'
-      >
-        <li className='dropdown-divider' role='presentation'>
-          <span className='filler'>&nbsp;</span>
-        </li>
-        {props.filterList.map((filterItem) => (
-          <li key={filterItem.value}>
-            <a
-              href='javascript:void(0)'
-              className='dropdown-item'
-              data-filter='grouping'
-              data-value={filterItem.value}
-              data-pref={filterItem.value}
-              aria-current={filterItem.value === props.filterValue}
-              onClick={() => props.setFilter(filterItem.value)}
-            >
-              {filterItem.display}
-            </a>
-          </li>
-        ))}
-      </ul>
-    </div>
-  </div>
-);
-
-const QuickCourseViewBody = (props: { courses: Course[] }) => (
-  <div className='container-fluid p-0'>
-    <div
-      data-region='courses-view'
-      data-display='list'
-      data-grouping='all'
-      data-customfieldname
-      data-customfieldvalue
-    >
-      <div data-region='paged-content-container'>
-        <div
-          data-region='page-container'
-          className='paged-content-page-container'
-          aria-live='polite'
-        >
-          <div data-region='paged-content-page' data-page='1'>
-            <ListGroup
-              items={props.courses.map((course) => ({
-                key: course.fullName,
-                element: <CourseItem course={course} />,
-              }))}
-            />
-          </div>
-        </div>
-      </div>
-      <div className='mt-1'></div>
-    </div>
-  </div>
-);
+import { Course, RegularLectureCourse } from '../../../common/storage/course.ts';
+import QuickCourseViewControl from './QuickCourseViewControl.tsx';
+import QuickCourseViewBody from './QuickCourseViewBody.tsx';
+import { semesterMap, semesterOrder, weekOfDayOrder, Filter } from './defs.ts';
 
 const filterCourse = function (courses: Course[], filter: string) {
   if (filter === 'all') {
